feat: Update Belief class and modify tests
Belief class was updated to change the access modifier of "_observation" to public. This property is now accessible via the public property "Observation". The tests were amended to account for this change, allowing checks on the 'Observation' property in the BeliefTests class. The folder reference to "Stubs" in Aplib.Core.Tests.csproj was also removed as it was unnecessary.